feat!: profile picture roundness
should close #377 if everything goes according to plan

Icons come from the stoatchat/for-android repository (converted from
android vectordrawable to svg)

This adds a new theme option, customRadii, that tells the avatar
component to change it's roundness (as a percentage). It also removes
the "shape" prop, as i don't think will be needed anymore (for now). To
override the roundness of the avatar, use the "roundness" prop instead.

Very breaking PR, btw

Signed-off-by: Amy <amy+git@amogus.cloud>